Adds day, beginning hour and ending hour field to room search
Adds extra fields to room search forms, as a means of separating the selection of the reservation day and it'd begin and end hours.
Also adds a couple of string formating methods to better support interactions with the MySQL database.
Finally, makes the actual reservation begin and end fields immutable directly, thus, they are fixed based on the values passed by POST during the room search.

Commit: Add

TaskNumber: -