Adds special reservation menu
Creates a couple of modals where the user can enter some details(time begin, time end and description) to create a requisition for a room in specific. This requisition is to be done for cases which the system, already in place, is unable to accomodate.
Needs to change or fix the Timepicker.

Commit: Add

TaskNumber: -