Initial solution for direct reservation bug
Adds a specific 'path' that resolves the problem of a direct reservation not having valid beginning and end dates. In said path, date and times are changeable just the same as in the research view. Also changes the calendar view to respect another business rule, that reservations can only be made within the current semester.

Commit: Add

TaskNumber: #14